+++ NEWS.md
@@ -10,6 +10,8 @@
 - Added VSCLSIG and RACE to ADVS
 - Added AOCTIFL, AESCAT, SMQzzNAM, and CQzzNAM to ADAE
 - Added derived variables `AESHOSPP` and `AESHOSPR`
+- Added variables `COHORT`, `GROUP`, and `EOTDT` to ADSL.
+- Added variables `CMSTRTPT` and `CMSTTPT` to ADCM.
 
 
 ### Changed